requirements.txt
aiogram==3.0.0
firebase-admin==6.2.0
apscheduler==3.10.4
python-dotenv==1.0.0
